fixes #22: Compatibility with Openfire 4.9.0 and later
In Openfire 4.9.0, a deprecated API was removed that was used by this plugin. This commit replaces that usage with usage of an API that was introduced in Openfire 4.8.0. As a result, this plugin now requires Openfire 4.8.0 or later.

As the plugin no longer is compatible with older versions of Openfire, a backwards-compatibily hack for JSP compilation was removed.